refactor(mobile): use the same advanced tx details view on confirmTX and in history
