+++ apps/mobile/src/app/_layout.tsx
@@ -118,10 +118,6 @@ function RootLayout() {
                               name="history-transaction-details"
                               options={{ headerShown: true, title: 'Transaction details' }}
                             />
-                            <Stack.Screen
-                              name="history-advanced-details"
-                              options={{ headerShown: true, title: 'Advanced Details' }}
-                            />
                             <Stack.Screen name="address-book" options={{ headerShown: true, title: '' }} />
                             <Stack.Screen name="contact" options={{ headerShown: true, title: '' }} />
                             <Stack.Screen name="signers" options={{ headerShown: false }} />
