+++ apps/mobile/src/features/ConfirmTx/components/confirmation-views/AddSigner/AddSigner.tsx
@@ -9,7 +9,7 @@ import { selectChainById } from '@/src/store/chains'
 
 import { ListTable } from '../../ListTable'
 import { TransactionHeader } from '../../TransactionHeader'
-import { ParametersButton } from '../../ParametersButton'
+import { ParametersButton } from '@/src/components/ParametersButton'
 import { NormalizedSettingsChangeTransaction } from '../../ConfirmationView/types'
 import { useOpenExplorer } from '@/src/features/ConfirmTx/hooks/useOpenExplorer'
