+++ apps/mobile/src/features/ConfirmTx/components/confirmation-views/Stake/Deposit/Deposit.tsx
@@ -9,7 +9,7 @@ import {
   TransactionData,
 } from '@safe-global/store/gateway/AUTO_GENERATED/transactions'
 import { TokenAmount } from '@/src/components/TokenAmount'
-import { ParametersButton } from '../../../ParametersButton'
+import { ParametersButton } from '@/src/components/ParametersButton'
 
 interface StakingDepositProps {
   txInfo: NativeStakingDepositTransactionInfo
