+++ apps/mobile/src/features/ConfirmTx/components/confirmation-views/VaultDeposit/VaultDeposit.tsx
@@ -9,7 +9,7 @@ import { TransactionHeader } from '../../TransactionHeader'
 import { ListTable } from '../../ListTable'
 import { TokenAmount } from '@/src/components/TokenAmount'
 import { formatPercentage } from '@safe-global/utils/utils/formatters'
-import { ParametersButton } from '../../ParametersButton'
+import { ParametersButton } from '@/src/components/ParametersButton'
 import { vaultTypeToLabel, formatVaultDepositItems } from './utils'
 import { Container } from '@/src/components/Container'
 import { Image } from 'expo-image'
