+++ apps/mobile/src/features/ConfirmTx/components/confirmation-views/VaultRedeem/VaultRedeem.tsx
@@ -8,7 +8,7 @@ import { TransactionHeader } from '../../TransactionHeader'
 import { ListTable } from '../../ListTable'
 import { TokenAmount } from '@/src/components/TokenAmount'
 import { formatPercentage } from '@safe-global/utils/utils/formatters'
-import { ParametersButton } from '../../ParametersButton'
+import { ParametersButton } from '@/src/components/ParametersButton'
 import { Container } from '@/src/components/Container'
 import { vaultTypeToLabel } from '../VaultDeposit/utils'
 import { formatVaultRedeemItems } from './utils'
