--- apps/mobile/src/features/HistoryAdvancedDetails/index.ts
@@ -1,1 +1,0 @@
-export { HistoryAdvancedDetailsContainer } from '@/src/features/HistoryAdvancedDetails/HistoryAdvancedDetails.container'
