+++ apps/mobile/src/features/HistoryTransactionDetails/components/history-views/HistoryStakeDeposit.tsx
@@ -12,7 +12,7 @@ import {
   formatStakingDepositItems,
   formatStakingValidatorItems,
 } from '@/src/features/ConfirmTx/components/confirmation-views/Stake/utils'
-import { HistoryAdvancedDetailsButton } from '../HistoryAdvancedDetailsButton'
+import { HistoryAdvancedDetailsButton } from '@/src/features/HistoryTransactionDetails/components/HistoryAdvancedDetailsButton'
 
 interface HistoryStakeDepositProps {
   txId: string
